+++ CHANGES.rst
@@ -6,6 +6,7 @@ Version 0.8.2 (UNRELEASED)
 
 - Adds email validation to the ``user-create`` command used by the REANA administrators.
 - Adds workflow name validation to the ``create_workflow`` endpoint.
+- Changes ``/api/info`` endpoint to return a list of supported compute backends.
 - Changes ``/api/status`` endpoint to calculate the cluster health status based on the availablity instead of the usage.
 
 Version 0.8.1 (2021-11-29)